feat: Enhance DGstPlayerProxy for Qt6 support
Add support for audio output and media error handling in DGstPlayerProxy when built with Qt6. Introduce a new slot for media errors and connect it to the QMediaPlayer's error signal. Update volume and mute functionalities to utilize QAudioOutput. Modify VideoSurface to use QVideoSink for video frame handling in Qt6.

Log: Enhance DGstPlayerProxy for Qt6 support